+++ Core/Models/Customer.cs
@@ -1,7 +1,7 @@
 ﻿namespace Core.Models;
 
-public class Customer(long id = 0, string name = "")
+public class Customer(int id = 0, string name = "")
 {
-    public long Id { get; set; } = id;
+    public int Id { get; set; } = id;
     public string Name { get; private set; } = name;
 }
