+++ DataContext/Models/Customer.cs
@@ -1,3 +1,3 @@
 namespace DataContext.Models;
 
-public record Customer(long Id = 0, string Name = "");
+public record Customer(int Id = 0, string Name = "");
